refactor: extract podcast detail page pagination tests
